requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
Markdown==3.5.2
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
